+++ feature/pro/availability/src/commonMain/kotlin/com/hulkdx/findprofessional/feature/pro/availability/main/model/ProAvailabilityState.kt
@@ -5,7 +5,7 @@ import kotlin.math.ceil
 data class ProAvailabilityState(
     val currentMonth: String,
     val firstDay: Int,
-    val lengthOfMonth: Int,
+    val lengthOfMonth: Long,
 ) {
     val perWeek = ceil((lengthOfMonth + firstDay) / 7F).toInt()
 }
